requirements.txt
attrs
Automat==20.2.0
cffi==1.14.0
constantly==15.1.0
cryptography==3.3.2
cssselect==1.1.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.6.3
numpy==1.18.2
pandas
#pandas==1.0.3
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
requests
psycopg2
Scrapy
selenium
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
urllib3==1.26.5
w3lib==1.21.0
zope.interface==5.0.0
redis
types-redis
python-dotenv
pyarrow
jsonlines
scrapy-redis